feat: add IdentityHub client (#4)
* Added identity hub client to get verifiable credentials (#8)

Added identity hub client to get verifiable credentials.


* Fixed checkstyle.

* Added description in changelog.

* Use var for consistency.

* Removed MessageStatus builder.

* Removed RequestStatus builder.

* Document possibility to expand to handle other types of objects and operations.

* Fixed unused imports.

* Moved dtos in dtos module.

* Return status result instead of throwing ApiException to get credentials.

* Changed contract of client .

* Small improvement.

* Renamed module dtos to identity-hub-dtos.

* Added : for consistency.
