+++ extensions/identity-hub/src/test/java/org/eclipse/dataspaceconnector/identityhub/store/IdentityHubInMemoryStoreTest.java
@@ -15,12 +15,11 @@
 package org.eclipse.dataspaceconnector.identityhub.store;
 
 import com.github.javafaker.Faker;
-import org.eclipse.dataspaceconnector.identityhub.api.VerifiableCredential;
+import org.eclipse.dataspaceconnector.identityhub.dtos.credentials.VerifiableCredential;
 import org.junit.jupiter.api.Test;
 
-import java.util.List;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
+import java.util.stream.IntStream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -32,8 +31,9 @@ public class IdentityHubInMemoryStoreTest {
         // Arrange
         var store = new IdentityHubInMemoryStore();
         var credentialsCount = FAKER.number().numberBetween(1, 10);
-        List<VerifiableCredential> credentials = Stream.generate(() -> VerifiableCredential.Builder.newInstance().id(FAKER.internet().uuid()).build())
-                .limit(credentialsCount).collect(Collectors.toList());
+        var credentials = IntStream.range(0, credentialsCount)
+                .mapToObj(i -> VerifiableCredential.Builder.newInstance().id(FAKER.internet().uuid()).build())
+                .collect(Collectors.toList());
 
         // Act
         credentials.forEach(store::add);
