+++ gradle.properties
@@ -5,8 +5,10 @@ jupiterVersion=5.8.2
 storageBlobVersion=12.11.0
 mockitoVersion=4.2.0
 awaitility=4.1.1
-rsApi=3.0.0
+rsApi=3.1.0
 swagger=2.1.13
 jacksonVersion=2.13.1
 restAssured=4.5.0
 faker=1.0.2
+okHttpVersion=4.9.3
+jetBrainsAnnotationsVersion=15.0
