+++ identity-hub-client/src/test/java/org/eclipse/dataspaceconnector/identityhub/client/IdentityHubClientImplIntegrationTest.java
@@ -1,0 +1,53 @@
+/*
+ *  Copyright (c) 2022 Microsoft Corporation
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Microsoft Corporation - initial API and implementation
+ *
+ */
+
+package org.eclipse.dataspaceconnector.identityhub.client;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.github.javafaker.Faker;
+import org.eclipse.dataspaceconnector.common.testfixtures.TestUtils;
+import org.eclipse.dataspaceconnector.identityhub.dtos.credentials.VerifiableCredential;
+import org.eclipse.dataspaceconnector.junit.launcher.EdcExtension;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+@ExtendWith(EdcExtension.class)
+public class IdentityHubClientImplIntegrationTest {
+
+    private static final String API_URL = "http://localhost:8181/api/identity-hub";
+    private static final Faker FAKER = new Faker();
+    private static final String VERIFIABLE_CREDENTIAL_ID = FAKER.internet().uuid();
+    private static final ObjectMapper OBJECT_MAPPER = new ObjectMapper();
+    private IdentityHubClient client;
+
+    @BeforeEach
+    void setUp() {
+        var okHttpClient = TestUtils.testOkHttpClient();
+        client = new IdentityHubClientImpl(okHttpClient, OBJECT_MAPPER);
+    }
+
+    @Test
+    void addAndQueryVerifiableCredentials() throws Exception {
+        var credential = VerifiableCredential.Builder.newInstance().id(VERIFIABLE_CREDENTIAL_ID).build();
+
+        client.addVerifiableCredential(API_URL, credential);
+        var statusResult = client.getVerifiableCredentials(API_URL);
+
+        assertThat(statusResult.succeeded());
+        assertThat(statusResult.getContent()).usingRecursiveFieldByFieldElementComparator().contains(credential);
+    }
+}
