+++ spi/identity-hub-store-spi/build.gradle.kts
@@ -1,0 +1,3 @@
+plugins {
+    `java-library`
+}
